Extract downloadPeerFile
