+++ lib/slsk-client.js
@@ -15,13 +15,13 @@ const debugPeerFile = require('debug')('slsk:peer:file')
 
 const Messages = require('./messages.js')
 const MessageFactory = require('./message-factory.js')
+const downloadPeerFile = require('./download-peer-file.js')
 let stack = require('./stack')
 
 // TODO make this to store, to be required by other modules
 let client
 let peers = {}
 let rest
-let currentLogin
 
 class SlskClient extends EventEmitter {
   init (cb) {
@@ -127,12 +127,7 @@ class SlskClient extends EventEmitter {
 
     if (type === 'F') {
       // debug(`connectToPeer ${user} ${host} ${port} ${token} ${type}`)
-      this.downloadPeerFile({
-        user,
-        host,
-        port,
-        token
-      })
+      downloadPeerFile(host, port, token, user, false)
       return
     }
 
@@ -233,15 +228,8 @@ class SlskClient extends EventEmitter {
               debugPeer(`reason: ${reason}, I will receive TransferRequest soon...`)
               delete stack.downloadTokens[token] // avoid memory leak
             } else if (allowed === 1) {
-              // let size = msg.int32()
               debugPeer(`Directly allowed. Connecting to peer with PeerInit + ${token}`)
-              // let down = stack.downloadTokens[token]
-              this.downloadPeerFile({
-                host,
-                port,
-                token,
-                user
-              }, true)
+              downloadPeerFile(host, port, token, user, true)
             }
             break
           }
@@ -266,92 +254,12 @@ class SlskClient extends EventEmitter {
   }
 
   login (credentials, cb) {
-    currentLogin = credentials.user
+    stack.currentLogin = credentials.user
     let msg = MessageFactory
       .to.server
       .login(credentials)
     client.write(msg.getBuff())
     stack.login = cb
-  }
-
-  downloadPeerFile ({host, port, token, user, stream}, noPierce) {
-    debugPeerFile(`downloadPeerFile ${user}`)
-    let conn = net.createConnection({
-      host,
-      port
-    }, () => {
-      if (noPierce) {
-        debugPeerFile(`noPierce ${user} connected`)
-        let m = MessageFactory.to.peer.peerInit(currentLogin)
-        let noPierceB = m.getBuff()
-        noPierceB = Buffer.concat([noPierceB, Buffer.from(token, 'hex')])
-        hex(noPierceB)
-        conn.write(noPierceB)
-        setTimeout(() => {
-          debugPeerFile(`noPierce sending 8* 00`)
-          conn.write(Buffer.from('00000000' + '00000000', 'hex'))
-        }, 1000)
-        return
-      }
-      let pierceFw = MessageFactory.to.peer.pierceFw(token)
-      conn.write(pierceFw.getBuff())
-    })
-
-    let received = false
-    let requestToken
-    if (noPierce) requestToken = token
-    let buf = Buffer.alloc(0)
-    let tok
-    let down
-    let i = 0
-    conn.on('data', data => {
-      // debugPeerFile(`data: ${data.length}`)
-      if (!noPierce && !received) {
-        requestToken = data.toString('hex', 0, 4)
-        conn.write(Buffer.from('00000000' + '00000000', 'hex'))
-        received = true
-      } else {
-        debug(`file data`)
-        if (down && down.stream) {
-          debug('push to stream')
-          down.stream.push(data)
-        }
-        buf = Buffer.concat([buf, data])
-      }
-      if (tok) {
-        if (i % 10 === 0) {
-          debugPeerFile(`buf: ${buf.length} size: ${tok.size}`)
-        }
-        i++
-      } else {
-        tok = stack.downloadTokens[requestToken]
-        down = stack.download[tok.user + '_' + tok.file]
-      }
-      if (tok && buf.length >= tok.size) {
-        debug(`disconnect, buf: ${buf.length} size: ${tok.size}`)
-        conn.end()
-      }
-    })
-    conn.on('close', () => {
-      debugPeerFile(`file socket close ${user}`)
-      if (tok && down) {
-        if (down.stream) down.stream.push(null)
-        const filePath = down.path || this.getFilePathName(tok.user, tok.file)
-        fs.writeFile(filePath, buf, () => {
-          down.path = filePath
-          down.buffer = buf
-          if (typeof down.cb === 'function') down.cb(null, down)
-        })
-      } else {
-        fs.writeFile('test.mp3', buf)
-        debugPeerFile(`ERROR: token ${token} not exist`)
-      }
-    })
-  }
-
-  getFilePathName (user, file) {
-    file = file.split('\\')
-    return `/tmp/slsk/${user}_${file[file.length - 1]}`
   }
 
   search (obj, cb) {
